refactored templating functionality to fit rebased code: routes in proper place, handlers handling properly
