+++ server/Mustache/template.html
@@ -8,7 +8,7 @@
 
 <body>
   <div>
-    <img src='{{albumart}}'></img>
+    <img src='{{album_art}}'></img>
     <h1> Play {{title}} <br> by {{artist}} on:</h1>
   </div>
