--- server/listener/listenerModel.js
@@ -1,10 +1,0 @@
-var mongoose = require('mongoose');
-
-var ListenerSchema = new mongoose.Schema({
-  title : String,
-  artist : String,
-  albumart: String,
-  providers: [{}]
-});
-
-module.exports = mongoose.model('Listener', ListenerSchema);
