+++ server/server.js
@@ -4,11 +4,9 @@ var path = require('path');
 var webpack = require('webpack');
 var webpackDevMiddleware = require('webpack-dev-middleware');
 var webpackHotMiddleware = require('webpack-hot-middleware');
-var mongoose = require('mongoose'); 
-
 
 var expressRouter = express.Router();
-var router = require('./router.js');
+var router = require('./routes/routes.js');
 
 var plainText = require('./requests/plainTextSearch.js');
 
@@ -16,7 +14,8 @@ var app = express();
 
 var port = 3000;
 var mongoose = require('mongoose');
-var mongoUrl = process.env.MONGOLAB_URI || 'mongodb://localhost/songlink';
+// var mongoUrl = process.env.MONGOLAB_URI || 'mongodb://localhost/songlink';
+var mongoUrl = 'mongodb://localhost/songlink';
 mongoose.connect(mongoUrl);
 
 var createHandler = require('./routes/createHandler');
